Ability to save multiple profiles for the same device
The current application saves configurations (like bot actions and base
layout) using the device "ip_port". However, in RoK, a lot of people
have multiple accounts in the same server and might need different
configurations for different accounts. For example, they may want to do
barbarians on the main account, but only gather resources in the farm
accounts. This is not possible as the entire device can save only one
configuration.

This commit adds the ability to save multiple configurations per device.
The user can add the same device multiple times with a new name and each
will store its configuration uniquely. The configuration files now use
the format "name_ip_port" to make each profile unique.

The user still needs to manually switch to the right account in RoK for
the profile selected - this is not done automatically.